split some utils methods
